+++ Sources/StreamChatSwiftUI/Images.swift
@@ -237,6 +237,7 @@ public class Images {
     public var play: UIImage = loadImageSafely(with: "play")
     public var playFilled: UIImage = UIImage(systemName: "play.fill")!
     public var pause: UIImage = loadImageSafely(with: "pause")
+    public var pauseFilled: UIImage = loadImageSafely(with: "pause.fill")
     
     public var checkmarkFilled: UIImage = UIImage(systemName: "checkmark.circle.fill")!